refactor: everything
This cleans up the stream objects and tests them. I'm sure there are some performance gains to be had.

I'm torn on the map, reduce, and filer streams and helper prototype methods. I had to ditch prototype methods because of issues with the circular dependencies but I'm not sure I really like them. I think it might be enough to have a `.promise()` on a stream an have it work with async functions.

The map, reduce, and filter streams themselves are pretty easy to reproduce, I'll need to practice using them and get people's thought to know if they're worth keeping.

todo:
- [x] docs
- [x] make everything a class
- [ ] write stream
- [x] basic tests for core streams
- [ ] make an async readBytes(x) bytes help function
- [x] make split object mode work in transform stream